Split threshold management into own module
